Move the processes state code over under the process package.
